GH #182. Scoped enumerators.

+++ Source/dooble_cookies.cc
@@ -112,7 +112,7 @@ void dooble_cookies::slot_cookie_added(const QNetworkCookie &cookie)
 {
   emit cookies_added
     (QList<QNetworkCookie> () << cookie,
-     QList<int> () << BlockedOrFavorite::NONE);
+     QList<int> () << static_cast<int> (BlockedOrFavorite::NONE));
 
   if(!dooble::s_cryptography || !dooble::s_cryptography->authenticated())
     return;
@@ -446,9 +446,9 @@ void dooble_cookies::slot_populate(void)
 	      cookie.setDomain(bytes);
 	      cookies << cookie;
 	      is_blocked_or_favorite <<
-		(is_blocked ? BlockedOrFavorite::BLOCKED :
-		 is_favorite ? BlockedOrFavorite::FAVORITE :
-		 BlockedOrFavorite::NONE);
+		static_cast<int> (is_blocked ? BlockedOrFavorite::BLOCKED :
+				  is_favorite ? BlockedOrFavorite::FAVORITE :
+				  BlockedOrFavorite::NONE);
 	    }
 
 	if(!cookies.isEmpty() && !is_blocked_or_favorite.isEmpty())
@@ -572,9 +572,9 @@ void dooble_cookies::slot_populate(void)
 	      c.setDomain(""); // Limit the cookie to the exact server.
 	      count += 1;
 	      is_blocked_or_favorite <<
-		(is_blocked ? BlockedOrFavorite::BLOCKED :
-		 is_favorite ? BlockedOrFavorite::FAVORITE :
-		 BlockedOrFavorite::NONE);
+		static_cast<int> (is_blocked ? BlockedOrFavorite::BLOCKED :
+				  is_favorite ? BlockedOrFavorite::FAVORITE :
+				  BlockedOrFavorite::NONE);
 	      profile->cookieStore()->setCookie(c, url);
 	    }
